Move Credential definition to root of crate
