+++ azure_devops_rust_api/src/lib.rs
@@ -85,4 +85,5 @@ pub mod wit;
 #[cfg(feature = "work")]
 pub mod work;
 
-pub mod auth;
+mod auth;
+pub use auth::Credential;
